Open Aries v0.1.19: FAT32 & ext2 full support, project branding
Branding:
  - Project name 'Open Aries' in boot banner and info command

FAT32 fixes:
  - open_file_t.start_cluster widened to uint32_t (was uint16_t)
  - Removed fd->offset hack for 32-bit cluster storage
  - Added resolve_fat32_path(), find_in_dir(), free_slot_in_dir()
  - All ops path-aware: create/open/delete/stat/mkdir/rmdir/rename
  - mkdir '..' points to actual parent, not hardcoded root
  - walk_dir: callback before end-of-dir check (fixes slot detection)

ext2 fixes:
  - ext2_rmdir uses resolve_path() instead of hardcoded root inode

Tooling:
  - tools/ext2_inject.py: inject files into ext2 images on macOS

All three filesystems (FAT16, FAT32, ext2) pass identical test suite:
  file create/append/cat/stat/wc/cp/rename/rm, nested mkdir/rmdir,
  pipes, I/O redirection, shell scripts, external program execution

+++ include/aios/vfs.h
@@ -18,7 +18,7 @@ typedef struct aios_fs_ops aios_fs_ops_t;
 
 typedef struct {
     int      in_use;
-    uint16_t start_cluster;
+    uint32_t start_cluster;
     uint32_t file_size;
     uint32_t offset;
     uint32_t dir_abs_sector; /* absolute sector holding this dir entry  */
